requirements.txt
apache-airflow==2.6.1
apache_airflow_providers_common_sql==1.3.4
python-dotenv==1.0.0
psycopg2==2.9.6
psycopg2_binary==2.9.6
requests==2.28.2
